Changed first argument of myax() to axes handle
Since its inception myax() has taken fig or p.gcf() as its first argument.
To avoid p.sca() calls when working with multi-axes figures, and to be more consistent with all the other functions in figfun, I have changed the first argument to an axes handle.  For the time being, both axes and figure are acceptable.